feat: a full program which runs correctly
