+++ src/gcc/example.c
@@ -11,44 +11,22 @@
 #include <stdio.h>
 #include <math.h> 
 #include <stdlib.h> 
-
-typedef struct MyNode
-{
-	int val;
-	struct MyNode* next;
-}Node;
-
-void addResult(Node** _p_result_head, Node** _p_result_tail, int _val)
-{
-	/*printf("check %d\n",_val);*/
-	Node* t = (Node*)malloc(sizeof(Node));
-	t->val = _val;
-	t->next = NULL;
-	//add the new node to the tail
-	if(*_p_result_head == NULL)
-	{
-		*_p_result_head = t;
-		*_p_result_tail = *_p_result_head;
-	}
-	else
-	{
-		(*_p_result_tail)->next = t;
-		*_p_result_tail = t;
-	}
-}
+#include <limits.h> 
+#include "list.h"
 
 //NOTICE: sqrt(8) works but sqrt(n) must be compiled by -lm
 int main(int argc, const char* argv[])
 {
+	//int list1[] = {1,3,5,8};
 	int list1[] = {1,3,5,7};
 	int list1_num = 4;
 	int list2[] = {2,4,6,8};
 	int list2_num = 4;
 
-	Node *result_head, *result_tail;
+	Node *result_head = NULL, *result_tail = NULL;
 	int pos1 = 0, pos2 = 0;
 	//NOTICE: this strategy not always fails, for example, the final elements of the two lists are the same
-	while(pos1 < list1_num || pos2 < list2_num)
+	while(pos1 < list1_num && pos2 < list2_num)
 	{
 		/*printf("check %d %d\n", pos1, pos2);*/
 		if(list1[pos1] < list2[pos2])
@@ -69,6 +47,18 @@ int main(int argc, const char* argv[])
 		}
 	}
 
+	while(pos1 < list1_num)
+	{
+		addResult(&result_head, &result_tail, list1[pos1]);
+		pos1++;
+	}
+	while(pos2 < list2_num)
+	{
+		addResult(&result_head, &result_tail, list2[pos2]);
+		pos2++;
+	}
+
+
 	//output each item in result after sqrt operation
 	Node* p = result_head;
 	while(p != NULL)
@@ -81,12 +71,13 @@ int main(int argc, const char* argv[])
 	//release the result list
 	p = result_head;
 	//NOTICE: another bug exists here, but error is not reported when running
-	printf("%p %p\n", result_head, result_tail);
+	//printf("%p %p\n", result_head, result_tail);
 	while(p != NULL)
 	{
+		Node* tmp = p->next;
 		free(p);
-		p = p->next;
-		printf("%p\n", p);
+		p = tmp;
+		//printf("%p\n", p);
 	}
 
 	return 0;
